+++ database/migrations/2021_10_22_060355_create_products_table.php
@@ -22,7 +22,7 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('short_description')->nullable();
             $table->text('description');
-            $table->bigInteger('price');
+            $table->double('price');
             $table->timestamps();
         });
     }
